third_party/blink/web_tests/fast/hidpi/2x/dpr-setting.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    addEventListener("load", function() {
        shouldBeTrue('matchMedia("screen and (min-resolution: 2dppx)").matches');
    }, false);
</script>
<div>This test passes if the DPR is set to 2x</div>
