third_party/blink/web_tests/fast/hidpi/2x/dpr-setting-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS matchMedia("screen and (min-resolution: 2dppx)").matches is true
This test passes if the DPR is set to 2x
